Move labrad types into separate module
