+++ labrad/servers/threaded_test_server.py
@@ -19,7 +19,8 @@ from __future__ import print_function
 
 import time
 
-from labrad import types as T, util
+from labrad.types import types as T
+from labrad import util
 from labrad.server import ThreadedServer, setting
 from labrad.units import m, s
 from labrad.util import hydrant
