+++ labrad/test/test_eval_lr_data.py
@@ -2,7 +2,7 @@ import math
 
 import pytest
 
-import labrad.types as T
+import labrad.types.types as T
 from labrad.units import Complex, Value
